test: move extra api server cert sans to its own package
